Add files via upload
This branch contains most of the initial studies found using search string: ("insect*" OR "coleopter*" OR "beetle*" OR "butterfl*" OR "moth*" OR "spider*") AND ("urban*") AND ("rural" OR "urban-rural comparison") AND ("life-history trait*" OR "fecund*" OR “reproduc*” OR “surviv*” OR “grow*” OR “dispers*”) NOT (“bird*” OR "child*" OR "mother*"). Many cells will be incomplete or missing, as this was used as a rough draft while deciding on a stringent effect size to use.